backend/requirements.txt
flask
flask-cors
transformers
torch
torchvision
pillow
python-dotenv
gunicorn
